+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- When within the same unit of time (seconds or milliseconds) 2 elapsed periods falsely reported than one was longer than the other.
+
 ## 4.1.0 - 2025-03-20
 
 ### Added